Move eth_call overrides setting to helper file
